moved function out of component into utlity
